pyproject.toml
[project]
name = "mcp-etcd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.9.1",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
]
